Refactor components to components directory
