+++ src/App.tsx
@@ -1,5 +1,4 @@
-import {useState, type FormEventHandler} from 'react';
-import {getArrivals, Bus} from '../backend/fetchArrivals';
+import {Arrivals} from '../components/Arrivals';
 
 function App() {
   return (
@@ -9,83 +8,6 @@ function App() {
         <Arrivals/>
     </>
   )
-}
-
-function Arrivals() {
-  const [arrivalsData, setArrivalsData] = useState<Bus[]>();
-  const [errorMessage, setErrorMessage] = useState<string>();
-  
-  async function handleGetArrivals(id: string){
-    let data = await getArrivals(id);
-    if( data.success){
-      setArrivalsData(data.array);
-    }
-    else{
-      setErrorMessage("No data found for stop code")
-    }
-  }
-
-  if (arrivalsData !== undefined){
-    return (
-      <>
-       <StopCodeForm onSubmit={handleGetArrivals}/>
-        <div>
-          {arrivalsData!.map((bus, index) => (
-            <BusCard busData={bus} key={index} />
-          ))}
-        </div>
-      </>
-    )}
-    else{
-      return (
-        <>
-          <h2>{errorMessage}</h2>
-          <StopCodeForm onSubmit={handleGetArrivals}/>
-        </>
-      );
-    }
-  }
-
-function StopCodeForm({onSubmit} : {onSubmit: Function}) {
-  const [stopCode, setStopCode] = useState<string>("");
-
-  const handleStopCodeChange = (event : React.ChangeEvent<HTMLInputElement>) => {
-    event.preventDefault();
-    setStopCode(event.currentTarget.value);
-  }
-
-  const handleSubmit = (event : React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault(); //prevent page frefresh
-    onSubmit(stopCode);
-  }
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <label> Enter a stop code:
-        <input 
-          type="text"
-          onChange={handleStopCodeChange}
-          value={stopCode}
-        />
-      </label>
-      <button type="submit" className="bg-blue-300 hover:bg-green-300 text-white font-semibold hover:text-white py-2 px-4 mb-3 border border-blue-500 hover:border-green-500 rounded">
-        Arrivals
-      </button>
-    </form>
-  )
-}
-
-function BusCard({busData} : {busData : Bus}){
-  if (busData !== null && busData.lineName !== ""){
-    let timeInMinues = Math.round(busData.timeToStation / 60);
-    return (
-    <>
-      <div className="max-w-sm rounded overflow-hidden shadow-lg">
-        <p>{busData.lineName} to {busData.towards} due in {timeInMinues} {timeInMinues === 1 ? "minute" : "minutes"}</p>
-      </div>
-    </>
-  )
-  }
 }
 
 export default App
